Some tidying up in code. Splitting modules. Cleaning commented lines and old unused things.
