+++ cases/structs.py
@@ -45,27 +45,19 @@ class CaseStructDef(SubCase):
 
     def split(self, elems:list[Elem])-> tuple[Expression, list[list[Elem]]]:
         typeName = elems[1].text
-        # prels('CaseStructDef.split', elems, show=1)
         # struct B(A,C)
         superNames = []
         subStart = 2
         if len(elems) > 2 and isLex(elems[2], Lt.oper, '('):
             # we have super-struct here
             brInd = firstBrackets(elems)
-            # print('brInd$1',brInd)
             brOpen, brClose = brInd
-            # print('Strc.split, brInd: ', brInd, elems[brOpen+1].text)
             superPart = elems[brOpen+1: brClose]
-            # print('SDF.split supers: {%s}' % elemStr(superPart))
             _, spl = CaseCommas().split(superPart)
-            # dprint('## spl:', [(n) for n in spl])
             superNames = [elemStr(n) for n in spl]
-            # dprint('## superNames:', superNames)
             subStart = brClose+1
-            
         
         sub = elems[subStart:]
-        # print('stru fields: {=%s=}'% elemStr(sub))
         exp = StructDefExpr(typeName, src=elems)
         exp.setSuper(superNames)
         subs = []
@@ -75,23 +67,10 @@ class CaseStructDef(SubCase):
         if cs.match(sub):
             _, subs = cs.split(sub)
         InterpretContext.get().addStruct(typeName)
-        # if typeName == 'B':
-            # raise XDebug('')
         return exp, subs
 
-    # def pairNorn(self, pair:ServPairExpr):
-    #     # if isinstance(pair.right, (ListConstr, DictConstr)):
-    #     #     # type parsed as block-constructor
-    #     #     match pair.right:
-    #     #         case ListConstr(): pair.right = VarExpr(VarUndefined('list'))
-    #     #         case DictConstr(): pair.right = VarExpr(VarUndefined('dict'))
-    #     return pair
-
     def setSub(self, base:StructDefExpr, subs:list[Expression])->Expression:
-        # print('CaseStructDef setSub', base, subs)
         for exp in subs:
-            # if isinstance(exp, ServPairExpr):
-            #     exp = self.pairNorn(exp)
             base.add(exp)
         return base
 
@@ -112,7 +91,6 @@ class CaseStructBlockDef(SubCase):
         return exp, subs
 
     def setSub(self, base:CaseStructDef, subs:Expression|list[Expression])->Expression:
-        # dprint('CaseStructBlockDef.setSub empty: ', base, subs)
         return base
 
 
@@ -140,7 +118,6 @@ class CaseStructConstr(SubCase, SolidCase):
         if len(elems) < 2:
             return False
         dc = CaseDictLine()
-        # print(dc)
         # get curvy brackets part
         # be sure that case already checked as Solid expr
         r =  isSolidExpr(elems, getLast=True)
@@ -151,8 +128,6 @@ class CaseStructConstr(SubCase, SolidCase):
         # print('lastFound:', ok, pos, 'lenEl:%d' % len(elems), elems[pos].text)
         if not ok or pos > len(elems)-2 or pos < 1 or not isLex(elems[pos], Lt.oper, '{') : 
             return False
-        # exit()
-        # print('Scon')
         return dc.match(elems[pos:])
     
     def split(self, elems:list[Elem])-> tuple[Expression, list[list[Elem]]]:
